Refactored accounts page to smaller components
